+++ lib/flutter_typeahead.dart
@@ -675,7 +675,8 @@ class _TypeAheadFieldState<T> extends State<TypeAheadField<T>>
   final Duration _resizeOnScrollRefreshRate = const Duration(milliseconds: 500);
 
   // Keyboard detection
-  KeyboardVisibilityNotification _keyboardVisibility = new KeyboardVisibilityNotification();
+  KeyboardVisibilityNotification _keyboardVisibility =
+      new KeyboardVisibilityNotification();
   int _keyboardVisibilityId;
 
   @override
@@ -719,8 +720,8 @@ class _TypeAheadFieldState<T> extends State<TypeAheadField<T>>
       },
     );
 
-    WidgetsBinding.instance.addPostFrameCallback((duration) async {
-      await this._initOverlayEntry();
+    WidgetsBinding.instance.addPostFrameCallback((duration) {
+      this._initOverlayEntry();
       // calculate initial suggestions list size
       this._suggestionsBoxController.resize();
 
@@ -758,10 +759,7 @@ class _TypeAheadFieldState<T> extends State<TypeAheadField<T>>
     });
   }
 
-  Future<void> _initOverlayEntry() async {
-    RenderBox renderBox = context.findRenderObject();
-    var size = renderBox.size;
-
+  void _initOverlayEntry() {
     this._suggestionsBoxController._overlayEntry =
         OverlayEntry(builder: (context) {
       final suggestionsList = _SuggestionsList<T>(
@@ -789,15 +787,16 @@ class _TypeAheadFieldState<T> extends State<TypeAheadField<T>>
       );
 
       return Positioned(
-        width: size.width,
+        width: _suggestionsBoxController.textBoxWidth,
         child: CompositedTransformFollower(
           link: this._layerLink,
           showWhenUnlinked: false,
           offset: Offset(
               0.0,
               widget.direction == AxisDirection.down
-                  ? size.height + widget.suggestionsBoxVerticalOffset
-                  : -widget.suggestionsBoxVerticalOffset),
+                  ? _suggestionsBoxController.textBoxHeight +
+                      widget.suggestionsBoxVerticalOffset
+                  : _suggestionsBoxController.directionUpOffset),
           child: widget.direction == AxisDirection.down
               ? suggestionsList
               : FractionalTranslation(
@@ -1369,7 +1368,6 @@ class TextFieldConfiguration<T> {
 }
 
 class _SuggestionsBoxController {
-  static const double defaultHeight = 300.0;
   static const int waitMetricsTimeoutMillis = 1000;
 
   final BuildContext context;
@@ -1379,7 +1377,10 @@ class _SuggestionsBoxController {
 
   bool _isOpened = false;
   bool widgetMounted = true;
-  double maxHeight = defaultHeight;
+  double maxHeight = 300.0;
+  double textBoxWidth = 100.0;
+  double textBoxHeight = 100.0;
+  double directionUpOffset;
 
   _SuggestionsBoxController(this.context, this.direction);
 
@@ -1446,11 +1447,14 @@ class _SuggestionsBoxController {
     if (widgetMounted) {
       TypeAheadField widget = context.widget as TypeAheadField;
 
+      RenderBox box = context.findRenderObject();
+      textBoxWidth = box.size.width;
+      textBoxHeight = box.size.height;
+
       if (direction == AxisDirection.down) {
         // height of window
         double h = MediaQuery.of(context).size.height;
 
-        RenderBox box = context.findRenderObject();
         // top of text box
         double textBoxAbsY = box.localToGlobal(Offset.zero).dy;
         double textBoxHeight = box.size.height;
@@ -1477,9 +1481,22 @@ class _SuggestionsBoxController {
             2 * widget.suggestionsBoxVerticalOffset;
       } else {
         // AxisDirection.up
-        RenderBox box = context.findRenderObject();
+        // Calculate offset when the keyboard is covering the suggestions box
+        // height of window
+        double h = MediaQuery.of(context).size.height;
+
+        // height of keyboard
+        double keyboardHeight = MediaQuery.of(context).viewInsets.bottom;
+
+        // recalculate keyboard absolute y value
+        double keyboardAbsY = h - keyboardHeight;
+
         // top of text box
         double textBoxAbsY = box.localToGlobal(Offset.zero).dy;
+
+        directionUpOffset = textBoxAbsY > keyboardAbsY
+            ? keyboardAbsY - textBoxAbsY - widget.suggestionsBoxVerticalOffset
+            : -widget.suggestionsBoxVerticalOffset;
 
         // we need to find the root MediaQuery for the unsafe area height
         // we cannot use BuildContext.ancestorWidgetOfExactType because
@@ -1489,9 +1506,13 @@ class _SuggestionsBoxController {
         // unsafe area, ie: iPhone X notch
         double unsafeAreaHeight = rootMediaQuery.data.padding.top;
 
-        maxHeight = textBoxAbsY -
-            unsafeAreaHeight -
-            2 * widget.suggestionsBoxVerticalOffset;
+        maxHeight = textBoxAbsY > keyboardAbsY
+            ? keyboardAbsY -
+                unsafeAreaHeight -
+                2 * widget.suggestionsBoxVerticalOffset
+            : textBoxAbsY -
+                unsafeAreaHeight -
+                2 * widget.suggestionsBoxVerticalOffset;
       }
 
       if (maxHeight < 0) maxHeight = 0;